Extract the classes and the constants
from "__init__.py".
