+++ pymcbdsc/constants.py
@@ -1,1 +1,4 @@
 version = "0.3.1"
+
+bds_version_pat = "[0-9]+\\.[0-9]+\\.[0-9]+\\.[0-9]+"
+bds_zip_file_pat = "bedrock-server-({version_pat})\\.zip".format(version_pat=bds_version_pat)
